Use E820 map to reserve BIOS and device memory
Keep it from being given to the memory allocator. Also get rid of init.pgalign
and use initmem area to allocate all the bootstrap data structures.

+++ include/arch-ia32/kimage.h
@@ -3,30 +3,30 @@
 
 #ifndef __ASM__
 
-extern uint32_t __begin;
+extern uint8_t __begin;
 
-extern uint32_t __text;
-extern uint32_t __text_end;
+extern uint8_t __text;
+extern uint8_t __text_end;
 
-extern uint32_t __rodata;
-extern uint32_t __rodata_end;
+extern uint8_t __rodata;
+extern uint8_t __rodata_end;
 
-extern uint32_t __data;
-extern uint32_t __data_end;
+extern uint8_t __data;
+extern uint8_t __data_end;
 
-extern uint32_t __bss;
-extern uint32_t __bss_end;
+extern uint8_t __bss;
+extern uint8_t __bss_end;
 
-extern uint32_t __kernel_end;
+extern uint8_t __kernel_end;
 
-extern uint32_t __init_start;
+extern uint8_t __init_start;
 
 extern initcall_t __initcall_start;
 extern initcall_t __initcall_end;
 
-extern uint32_t __init_end;
+extern uint8_t __init_end;
 
-extern uint32_t __end;
+extern uint8_t __end;
 
 #endif /* __ASM__ */
